+++ packages/sitecore-jss-dev-tools/src/disconnected-server/create-default-disconnected-server.ts
@@ -9,7 +9,6 @@ import {
   CustomizeRouteFunction,
 } from './DisconnectedLayoutServiceOptions';
 import { createDisconnectedLayoutService } from './layout-service';
-import { createDisconnectedSitemapService } from './sitemap-service';
 
 export interface DisconnectedServerOptions {
   appName: string;
@@ -150,12 +149,6 @@ export function createDefaultDisconnectedServer(options: DisconnectedServerOptio
         manifestLanguageChangeCallback: manifestManager.getManifest,
       });
 
-      // creates a fake version of the Sitecore Sitemap Service that is powered by your disconnected manifest file
-      const sitemapService = createDisconnectedSitemapService({
-        manifest,
-        getManifest: manifestManager.getManifest,
-      });
-
       // set up live reloading of the manifest when any manifest source file is changed
       manifestManager.setManifestUpdatedCallback((newManifest) => {
         layoutService.updateManifest(newManifest);
@@ -169,7 +162,6 @@ export function createDefaultDisconnectedServer(options: DisconnectedServerOptio
       app.use('/assets', Express.static(join(options.appRoot, 'assets')));
       app.use('/data/media', Express.static(join(options.appRoot, 'data/media')));
       app.use('/sitecore/api/layout/render', layoutService.middleware);
-      app.use('/sitecore/api/sitemap', sitemapService.middleware);
       app.use('/sitecore/api/jss/dictionary/:appName/:language', dictionaryService.middleware);
 
       if (options.afterMiddlewareRegistered) {
