+++ packages/sitecore-jss-nextjs/package.json
@@ -63,6 +63,7 @@
   },
   "dependencies": {
     "@sitecore-jss/sitecore-jss": "^16.0.0-canary.34",
+    "@sitecore-jss/sitecore-jss-manifest": "^16.0.0-canary.34",
     "@sitecore-jss/sitecore-jss-react": "^16.0.0-canary.34",
     "prop-types": "^15.7.2",
     "sync-disk-cache": "^2.1.0"
