+++ samples/nextjs/src/pages/[[...path]].tsx
@@ -11,7 +11,7 @@ import { StyleguideSitecoreContextValue } from 'lib/component-props';
 import { SitecorePageProps } from 'lib/page-props';
 import { sitecorePagePropsFactory } from 'lib/page-props-factory';
 import { componentFactory } from 'temp/componentFactory';
-import { sitemapFactory } from 'lib/sitemap-factory';
+import { sitemapFetcher } from 'lib/sitemap-fetcher';
 
 const SitecorePage = ({ notFound, layoutData, componentProps }: SitecorePageProps): JSX.Element => {
   useEffect(() => {
@@ -58,11 +58,11 @@ export const getStaticPaths: GetStaticPaths = async (context) => {
 
   if (process.env.NODE_ENV !== 'development') {
     // Note: Next.js runs export in production mode
-    const paths = await sitemapFactory.create(context);
+    const paths = await sitemapFetcher.fetch(context);
 
     return {
       paths,
-      fallback: false,
+      fallback: process.env.EXPORT_MODE ? false : 'blocking',
     };
   }
